62 proper noise terrain generation

+++ game/src/net/protocol.hpp
@@ -124,8 +124,8 @@ public:
 class NetTerrainMod final {
 public:
 	uint16_t x, y, w, h;
-	std::vector<uint8_t> tiles;
-	std::vector<int8_t> hmap;
+	std::vector<uint16_t> tiles;
+	std::vector<uint8_t> hmap;
 
 	static constexpr size_t possize = 4 * sizeof(uint16_t);
